Pass mode to GeneContents
